// src/styles/interfaces.ts
export interface GetSvgIconProps {
  width: number;
  height: number;
  iconName: string;
  color: string;
}
export interface TextInputWithIcon {
  color: string;
  focusColor: string;
}
export interface ColorChangeHover {
  initColor: string;
  changeColor: string;
  initBgColor: string;
  changeBgColor: string;
}